Login (#19)
* login and logout

* security

+++ taskmate/todolist_app/views.py
@@ -4,6 +4,7 @@ from todolist_app.models import TaskList
 from todolist_app.forms import TaskForm
 from django.contrib import messages
 from django.core.paginator import Paginator
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -13,20 +14,24 @@ def index(request):
     }
     return render(request, 'index1.html', context)
 
+@login_required
 def todolist(request):
     if request.method=='POST':
         form = TaskForm(request.POST or None)
         if form.is_valid():
-            form.save()
+            instance = form.save(commit=False)
+            instance.manage = request.user
+            instance.save()
         messages.success(request,('Task has beeen created'))
         return redirect('todolist')
     else:
-        all_tasks = TaskList.objects.all()
+        all_tasks = TaskList.objects.filter(manage=request.user)
         paginator = Paginator(all_tasks, 5)
         page = request.GET.get('pg')
         all_tasks = paginator.get_page(page)
         return render(request, 'todolist.html', {'all_tasks':all_tasks})
 
+@login_required
 def contact(request):
     context = {
         "contact_text":"Redirect to Google maps and type Kailasa you will find me.",
@@ -35,16 +40,20 @@ def contact(request):
 
 def about(request):
     context = {
-        "about_text":"I am god ",
+        "about_text":"I AM GOD ",
     }
     return render(request, 'about.html', context)
 
+@login_required
 def delete_task(request, item_id):
-    task = TaskList.objects.get(pk=item_id)
-    task.delete()
+    if task.manage == request.user:
+        task = TaskList.objects.get(pk=item_id)
+        task.delete()
+    else:
+        messages.error(request,('Permission Denied'))
     return redirect('todolist')
 
-
+@login_required
 def edit_task(request, item_id):
     if request.method =='POST':
         instance = get_object_or_404(TaskList, id=item_id)
@@ -57,9 +66,12 @@ def edit_task(request, item_id):
     else:
         task = TaskList.objects.get(pk=item_id)
         return render(request, 'edit.html', {'task':task})
-
+@login_required
 def update_status(request, item_id, status_flag):
     task = TaskList.objects.get(pk=item_id)
-    task.done = status_flag
-    task.save()
+    if task.manage == request.user:
+        task.done = status_flag
+        task.save()
+    else:
+        messages.error(request,('Permission Denied'))
     return redirect('todolist')
